requirements.txt
Brotli==1.0.7
click==7.1.2
dash==1.13.3
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.0
dash-table==4.8.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
patsy==0.5.1
plotly==4.8.1
plotly-express==0.4.1
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
scipy==1.4.1
six==1.15.0
statsmodels==0.11.1
Werkzeug==1.0.1

